+++ src/bench/types.d.ts
@@ -1,0 +1,3 @@
+declare module "js-sizeof" {
+    export default function sizeof(obj: unknown): number;
+}